refactor(mana_rich_text_ctrl): split rich text ctrl into mixin package
Move widgets/panels/mana_rich_text_ctrl.py into a package: frame.py
holds the border constants and both _ManaRichTextInner and
ManaSymbolRichCtrl class definitions; properties.py exposes GetValue
and content-detection helpers for each; handlers.py owns the setters,
rerendering, hint overlay, key handling, and painted-border layout.
