+++ widgets/panels/mana_rich_text_ctrl/__init__.py
@@ -1,0 +1,7 @@
+"""Mana-symbol-aware TextCtrl-compatible widget package."""
+
+from __future__ import annotations
+
+from widgets.panels.mana_rich_text_ctrl.frame import ManaSymbolRichCtrl
+
+__all__ = ["ManaSymbolRichCtrl"]
